Move MutationDetectors to runners/direct-java and tighten access
