+++ runners/direct-java/src/main/java/org/apache/beam/runners/direct/MutationDetectorsTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util;
+package org.apache.beam.runners.direct;
 
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.Lists;
@@ -24,10 +24,13 @@ import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
+import org.apache.beam.runners.direct.MutationDetector;
+import org.apache.beam.runners.direct.MutationDetectors;
 import org.apache.beam.sdk.coders.ByteArrayCoder;
 import org.apache.beam.sdk.coders.IterableCoder;
 import org.apache.beam.sdk.coders.ListCoder;
 import org.apache.beam.sdk.coders.VarIntCoder;
+import org.apache.beam.sdk.util.IllegalMutationException;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
